Improve default archive layout
